+++ src/refactoring/components/admin/coupon/CouponManagementSection.tsx
@@ -1,8 +1,8 @@
-import { useState } from "react";
 import { Coupon } from "../../../../types";
 import { SectionLayout } from "../../common";
 import { CouponForm } from "./CouponForm";
 import { CouponItemList } from "./CouponItemList";
+import { useCouponForm } from "../../../hooks";
 
 interface CouponManagementSectionProps {
   coupons: Coupon[];
@@ -13,26 +13,9 @@ export const CouponManagementSection = ({
   coupons,
   onCouponAdd,
 }: CouponManagementSectionProps) => {
-  const [newCoupon, setNewCoupon] = useState<Coupon>({
-    name: "",
-    code: "",
-    discountType: "percentage",
-    discountValue: 0,
+  const { newCoupon, handleAddCoupon, handleCouponChange } = useCouponForm({
+    onCouponAdd,
   });
-
-  const handleAddCoupon = () => {
-    onCouponAdd(newCoupon);
-    setNewCoupon({
-      name: "",
-      code: "",
-      discountType: "percentage",
-      discountValue: 0,
-    });
-  };
-
-  const handleCouponChange = (coupon: Coupon) => {
-    setNewCoupon(coupon);
-  };
 
   return (
     <SectionLayout title="쿠폰 관리">
